iidi/k8s/bc/panaroma-docker/requirements.txt
nats-py==2.2.0
requests==2.31.0